requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-import-export==2.8.0
